Add files via upload
A machine learning model that automatically recognizes handwritten digits (0–9) using image data. It is trained on the MNIST dataset, which contains thousands of grayscale images of handwritten numbers. The model uses a neural network (like a CNN) to learn patterns and classify new digit images with high accuracy. This project demonstrates the power of deep learning in image recognition tasks.